src/requirements.txt
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
loguru==0.7.2
minio==7.1.17
piexif==1.1.3
pika==1.3.2
Pillow==10.0.1
proglog==0.1.10
pyexiv2==2.5.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.7
